pyproject.toml
[project]
name = "dsg"
version = "0.1.0"
description = "Data site generator - a static site generator for data apps"
readme = "README.md"
authors = [
    { name = "Markus Hutnik", email = "markus.hutnik@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.1.3",
    "jinja2>=3.1.5",
    "markdown>=3.7",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "polars>=1.21.0",
    "pyarrow>=19.0.0",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]

[project.scripts]
dsg = "dsg:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
